refactor: make ring ingestion and persistence decisions testable
Two files carried most of this project's risk and were compiled by no test:
Calibration.cpp holds the only wire site for the snap-vs-slew invariant, and
Configuration.cpp holds every persistence decision. Neither can be added to
the harness directly - the test TU carries openvr_driver.h and Calibration.h
carries openvr.h, and the two cannot coexist. So the pure halves move into
headers the harness already compiles, and what genuinely needs the registry
or the OpenVR session stays put and is named as unverified.

Ring ingestion: the raw gate plus the compose-and-check pair appeared at four
call sites, so the overlay-side trust boundary the contract names was four
copies that had to stay in step, and the composed sample every solver input
depends on had no coverage at all despite being pure Eigen. It is now
TryComposeRingSample in RingPoseMath.h, one call per consumer, with four
scenarios. The three composed-time monotonicity guards are deliberately NOT
merged - buffer-relative, device-relative and observation-relative are
different domain rules, and each now says which it is.

Persistence: the write gates and the load-plan state machine are pure
functions over explicit inputs in ProfileValidation.h, and the profile record
and its JSON in a new ProfileRecordJson.h. Configuration.cpp keeps registry
I/O, context marshalling and the chaperone record, and shrinks by ~540 lines.
Save-side and load-side validation were two copies in two languages free to
drift; they are one function now, so a relaxed save bound also changes what
loads. Configuration.h is unchanged, so no caller moved.

Also: continuousState holds the enum rather than an int mirror of it, and
Clear() resets the two UI-visible counters it silently skipped. baseGeneration
deliberately still survives Clear - resetting it would let a recalibration
reuse a generation the driver already applied.

Vendored Eigen trips two more C++17 deprecations; silenced specifically
rather than with the blanket macro, which would also hide our own.

+++ Overlay/QuestCalibrator.cpp
@@ -679,7 +679,7 @@ static void SetupPreviewState()
 	CalCtx.mountExtrinsic.pos = Eigen::Vector3d(0.05, -0.08, 0.03);
 	CalCtx.mountExtrinsic.rotRmsDeg = 0.21;
 	CalCtx.mountExtrinsic.posRmsM = 0.004;
-	CalCtx.continuousState = static_cast<int>(questcal::ContinuousAlignment::State::Tracking);
+	CalCtx.continuousState = questcal::ContinuousAlignment::State::Tracking;
 	CalCtx.continuousDeviation.valid = true;
 	CalCtx.continuousDeviation.yawDeg = 0.08;
 	CalCtx.continuousDeviation.tiltDeg = 0.11;
